+++ frontend/App.tsx
@@ -2,20 +2,20 @@ import { StatusBar } from 'expo-status-bar';
 import { StyleSheet, View } from 'react-native';
 import { CommonActions, NavigationContainer } from '@react-navigation/native';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
-import Landing from './src/Landing';
-import Login from './src/Login';
-import Signup from './src/Signup';
-import Home from './src/Home';
-import Group from './src/Group';
-import CreateGroup from './src/CreateGroup';
-import Groups from './src/Groups';
-import CreateEvent from './src/CreateEvent';
+import Landing from './src/Landing/Landing';
+import Login from './src/Login/Login';
+import Signup from './src/Signup/Signup';
+import Home from './src/Home/Home';
+import Group from './src/Group/Group';
+import CreateGroup from './src/CreateGroup/CreateGroup';
+import Groups from './src/Groups/Groups';
+import CreateEvent from './src/CreateEvent/CreateEvent';
 import { Appbar, Provider as PaperProvider } from 'react-native-paper';
 
 import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
 import { Text, BottomNavigation } from 'react-native-paper';
 import Icon from 'react-native-vector-icons/MaterialCommunityIcons';
-import SearchGroups from './src/SearchGroups';
+import SearchGroups from './src/SearchGroups/SearchGroups';
 
 const Stack = createNativeStackNavigator();
 const Tab = createBottomTabNavigator();
