+++ frontend/src/SearchGroups/SearchGroups.tsx
